requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
nltk==3.9.1
numpy==2.1.2
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tqdm==4.66.5
tzdata==2024.2
python-dotenv==1.0.1